feat(bfan_importer): add bfan.link release importer
- adds a userscript for importing bfan.link smart-link releases through Harmony and adding missing provider relationships to MusicBrainz.
- extracts the reusable FFM importer workflow and provider URL logic into a shared smart-link importer library. Refactors the FFM importer to use the shared implementation.
- parses bfan.link provider destinations from its Next.js hydration data, ignores search-only fallbacks, and normalizes Apple Music album URLs. Adds parser and normalization tests, Vitest alias resolution, metadata, and documentation.
