+++ src/userscripts/bfan_importer/meta.json
@@ -1,0 +1,12 @@
+{
+    "name": "Import bfan.link releases to MusicBrainz",
+    "description": "Import bfan.link smart links with Harmony and add their remaining URL relationships to MusicBrainz",
+    "version": "2026.08.14.1",
+    "author": "Raman Sinclair",
+    "namespace": "https://github.com/murdos/musicbrainz-userscripts/",
+    "downloadURL": "https://raw.githubusercontent.com/murdos/musicbrainz-userscripts/dist/bfan_importer.user.js",
+    "updateURL": "https://raw.githubusercontent.com/murdos/musicbrainz-userscripts/dist/bfan_importer.user.js",
+    "match": ["https://bfan.link/*", "https://*.bfan.link/*"],
+    "runAt": "document-idle",
+    "icon": "https://raw.githubusercontent.com/murdos/musicbrainz-userscripts/master/assets/images/Musicbrainz_import_logo.png"
+}
