+++ vitest.config.ts
@@ -1,6 +1,13 @@
+import { fileURLToPath } from 'node:url';
+
 import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
+    resolve: {
+        alias: {
+            '~': fileURLToPath(new URL('./src', import.meta.url)),
+        },
+    },
     test: {
         environment: 'node',
         include: ['tests/**/*.test.ts'],
